Move sidebar button into header
