+++ client/src/styled-components/HeaderStyles.js
@@ -1,8 +1,8 @@
 import {styled} from '@mui/material/styles';
 import Box from '@mui/material/Box';
+import Button from '@mui/material/Button';
 
 export const HeaderLeftBox = styled(Box)(() => ({
-    marginLeft: '2.5em',
     fontSize: '1.5em', 
     fontWeight: 'bold',
     textAlign: 'center', 
@@ -13,4 +13,13 @@ export const HeaderRightBox = styled(Box)(() => ({
     width: 'fit-content',
     right: '1em',
     position: 'absolute'
+}));
+
+export const SidebarButton = styled(Button)(() => ({
+    zIndex: '1200',
+    margin: '0 8px',
+    padding: '0',
+    fontSize: '25px',
+    fontWeight: '700',
+    color: '#888',
 }));
