+++ client/src/styled-components/PageLayoutStyles.js
@@ -2,7 +2,6 @@ import {styled} from '@mui/material/styles';
 import {theme} from '../components/theme.jsx';
 import Box from '@mui/material/Box';
 import Drawer from '@mui/material/Drawer';
-import Button from '@mui/material/Button';
 
 const sidebarWidth = '400px';
 const sidebarWidthMobile = '75vw';
@@ -18,16 +17,6 @@ export const Sidebar = styled(Drawer)((props) => ({
     },
 }));
 
-export const SidebarButton = styled(Button)(() => ({
-    zIndex: '1200',
-    position: 'fixed',
-    fontSize: '25px',
-    fontWeight: '700',
-    color: '#888',
-    padding: '0',
-    height: '64px'
-}));
-
 export const ContentBox = styled(Box)((props) => ({
     transition: 'margin-left 225ms cubic-bezier(0, 0, 0.2, 1)',
     marginLeft: props.open && !props.mobile ? sidebarWidth : '0'
@@ -35,12 +24,4 @@ export const ContentBox = styled(Box)((props) => ({
 
 export const PageContent = styled(Box)((props) => ({
     margin: props.mobile ? '0 20px' : '0 60px'
-}));
-
-export const Header = styled(Box)(() => ({
-    height: '50px',
-    backgroundColor: theme.palette.background.page,
-    position: 'sticky',
-    top: '0',
-    width: '100%'
 }));
